Final tidy up
